+++ app/src/main/java/dev/mobile/medicalink/fragments/MainFragment.kt
@@ -12,12 +12,9 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentTransaction
 import dev.mobile.medicalink.R
-import dev.mobile.medicalink.db.local.AppDatabase
-import dev.mobile.medicalink.db.local.repository.MedocRepository
-import dev.mobile.medicalink.db.local.repository.UserRepository
 import dev.mobile.medicalink.fragments.home.HomeFragment
 import dev.mobile.medicalink.fragments.traitements.MainTraitementsFragment
-import dev.mobile.medicalink.fragments.traitements.MessagesFragment
+import dev.mobile.medicalink.fragments.messages.MessagesFragment
 
 /**
  * Fragment de la barre de navigation en bas de l'application (Accueil/Traitement/Messages)
