+++ app/src/main/java/dev/mobile/medicalink/fragments/traitements/ListeTraitementsFragment.kt
@@ -18,6 +18,8 @@ import dev.mobile.medicalink.db.local.AppDatabase
 import dev.mobile.medicalink.db.local.entity.Medoc
 import dev.mobile.medicalink.db.local.repository.MedocRepository
 import dev.mobile.medicalink.db.local.repository.UserRepository
+import dev.mobile.medicalink.fragments.traitements.adapter.ListeTraitementAdapterR
+import dev.mobile.medicalink.fragments.traitements.ajouts.AjoutManuelRecapitulatif
 import dev.mobile.medicalink.utils.notification.NotificationService
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
@@ -136,7 +138,6 @@ class ListeTraitementsFragment : Fragment() {
                     Pair(date, numero)
                 )
             }
-
         }
