+++ app/src/main/java/dev/mobile/medicalink/fragments/traitements/PreviewFragment.kt
@@ -23,6 +23,7 @@ import com.google.mlkit.vision.text.TextRecognition
 import com.google.mlkit.vision.text.TextRecognizer
 import com.google.mlkit.vision.text.latin.TextRecognizerOptions
 import dev.mobile.medicalink.R
+import dev.mobile.medicalink.fragments.traitements.ajouts.AddTraitementsFragment
 import java.io.File
 import java.io.InputStream
 import java.text.SimpleDateFormat
