+++ app/src/main/java/dev/mobile/medicalink/fragments/traitements/adapter/AjoutManuelTypeMedicAdapterR.kt
@@ -1,4 +1,4 @@
-package dev.mobile.medicalink.fragments.traitements
+package dev.mobile.medicalink.fragments.traitements.adapter
 
 import android.os.Build
 import android.util.Log
