+++ app/src/main/java/dev/mobile/medicalink/fragments/traitements/adapter/ListeEffetsSecondairesAdapterR.kt
@@ -1,4 +1,4 @@
-package dev.mobile.medicalink.fragments.traitements
+package dev.mobile.medicalink.fragments.traitements.adapter
 
 import android.os.Build
 import android.util.Log
@@ -9,6 +9,7 @@ import android.widget.TextView
 import androidx.annotation.RequiresApi
 import androidx.recyclerview.widget.RecyclerView
 import dev.mobile.medicalink.R
+import dev.mobile.medicalink.fragments.traitements.Traitement
 
 
 class ListeEffetsSecondairesAdapterR(private val list: MutableList<Traitement>) :
