+++ app/src/main/java/dev/mobile/medicalink/fragments/traitements/ajouts/AjoutManuelIntervalleRegulier.kt
@@ -1,4 +1,4 @@
-package dev.mobile.medicalink.fragments.traitements
+package dev.mobile.medicalink.fragments.traitements.ajouts
 
 import android.app.AlertDialog
 import android.content.Context
@@ -18,6 +18,7 @@ import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.fragment.app.Fragment
 import com.google.android.material.textfield.TextInputEditText
 import dev.mobile.medicalink.R
+import dev.mobile.medicalink.fragments.traitements.Traitement
 
 
 class AjoutManuelIntervalleRegulier : Fragment() {
