+++ app/src/main/java/dev/mobile/medicalink/fragments/traitements/ajouts/AjoutManuelSchemaPriseFragment.kt
@@ -1,4 +1,4 @@
-package dev.mobile.medicalink.fragments.traitements
+package dev.mobile.medicalink.fragments.traitements.ajouts
 
 import android.os.Build
 import android.os.Bundle
@@ -17,6 +17,8 @@ import androidx.annotation.RequiresApi
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.fragment.app.Fragment
 import dev.mobile.medicalink.R
+import dev.mobile.medicalink.fragments.traitements.Traitement
+import dev.mobile.medicalink.fragments.traitements.adapter.AjoutManuelTypeMedicAdapterR
 
 
 class AjoutManuelSchemaPriseFragment : Fragment() {
