+++ app/src/main/java/dev/mobile/medicalink/fragments/traitements/ajouts/AjoutManuelTypeMedic.kt
@@ -1,4 +1,4 @@
-package dev.mobile.medicalink.fragments.traitements
+package dev.mobile.medicalink.fragments.traitements.ajouts
 
 import android.os.Build
 import android.os.Bundle
@@ -15,6 +15,9 @@ import androidx.fragment.app.Fragment
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import dev.mobile.medicalink.R
+import dev.mobile.medicalink.fragments.traitements.SpacingRecyclerView
+import dev.mobile.medicalink.fragments.traitements.Traitement
+import dev.mobile.medicalink.fragments.traitements.adapter.AjoutManuelTypeMedicAdapterR
 
 
 class AjoutManuelTypeMedic : Fragment() {
