+++ app/src/main/res/layout/fragment_ajout_manuel_schema_prise2.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraint_layout_ajout_shema_prise_2"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.traitements.AjoutManuelSearchFragment">
+    tools:context=".fragments.traitements.ajouts.AjoutManuelSearchFragment">
 
     <TextView
         android:id="@+id/textView9"
