+++ app/src/main/res/layout/fragment_ajout_manuel_type_medic.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".fragments.traitements.AjoutManuelSearchFragment">
+    tools:context=".fragments.traitements.ajouts.AjoutManuelSearchFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewTypeMedic"
